Labo_02_Introductie_MVVM

+++ _Herhaling/Labo_02_Introductie_MVVM/NMCT.TipCalculator/NMCT.TipCalculator.Core/Services/CalculationService.cs
@@ -8,7 +8,7 @@ namespace NMCT.TipCalculator.Core.Services
 {
     public class CalculationService : ICalculationService
     {
-        public decimal Calculate(decimal payed, decimal procent)
+        public double Calculate(double payed, double procent)
         {
             return (payed * procent) / 100;
         }
